Add argument types to ITIR when tracing

+++ tests/next_tests/integration_tests/feature_tests/iterator_tests/test_scan.py
@@ -33,7 +33,7 @@ def test_scan_in_stencil(program_processor, lift_mode):
     ksize = 3
     Koff = offset("Koff")
     inp = gtx.np_as_located_field(IDim, KDim)(
-        np.copy(np.broadcast_to(np.arange(0, ksize), (isize, ksize)))
+        np.copy(np.broadcast_to(np.arange(0, ksize, dtype=np.float64), (isize, ksize)))
     )
     out = gtx.np_as_located_field(IDim, KDim)(np.zeros((isize, ksize)))
